moved plotting script into wfl.fit.error
